Extract tests for breakpoints in tests/debugging/test_breakpoints.py
