+++ numba_dppy/tests/debugging/test_debug_dppy_numba.py
@@ -58,39 +58,6 @@ def test_break_conditional(app):
     app.child.expect(r"\$1 = 1")
 
 
-@skip_no_numba055
-@pytest.mark.parametrize(
-    "breakpoint, api",
-    [
-        ("side-by-side.py:25", "numba"),
-        ("side-by-side.py:25", "numba-dppy-kernel"),
-        ("common_loop_body_242", "numba"),
-        ("common_loop_body", "numba-dppy-kernel"),
-    ],
-)
-def test_breakpoint_with_condition_by_function_argument(app, breakpoint, api):
-    """Function breakpoints and argument initializing
-
-    Test that it is possible to set conditional breakpoint at the beginning
-    of the function and use a function argument in the condition.
-
-    Test for https://github.com/numba/numba/issues/7415
-    SAT-4449
-    """
-    variable_name = "param_a"
-    variable_value = "3"
-    condition = f"{variable_name} == {variable_value}"
-
-    app.breakpoint(f"{breakpoint} if {condition}")
-    app.run(f"side-by-side.py --api={api}")
-
-    app.child.expect(r"Thread .* hit Breakpoint .* at side-by-side.py:25")
-
-    app.print(variable_name)
-
-    app.child.expect(fr"\$1 = {variable_value}")
-
-
 # commands/break_file_func
 def test_break_file_function(app):
     app.breakpoint("simple_sum.py:data_parallel_sum")
